+++ phtree/benchmark/update_box_d_benchmark.cc
@@ -24,11 +24,13 @@ using namespace improbable::phtree::phbenchmark;
 
 namespace {
 
-constexpr int UPDATES_PER_ROUND = 1000;
+constexpr size_t UPDATES_PER_ROUND = 1000;
 constexpr double MOVE_DISTANCE = 10;
 
 const double GLOBAL_MAX = 10000;
 const double BOX_LEN = 10;
+
+enum UpdateType { RELOCATE, ERASE_BY_KEY };
 
 template <dimension_t DIM>
 using BoxType = PhBoxD<DIM>;
@@ -46,14 +48,12 @@ struct UpdateOp {
 /*
  * Benchmark for updating the position of entries.
  */
-template <dimension_t DIM>
+template <dimension_t DIM, UpdateType UPDATE_TYPE>
 class IndexBenchmark {
   public:
     IndexBenchmark(
         benchmark::State& state,
-        TestGenerator data_type,
-        int num_entities,
-        int updates_per_round = UPDATES_PER_ROUND,
+        size_t updates_per_round = UPDATES_PER_ROUND,
         double move_distance = MOVE_DISTANCE);
 
     void Benchmark(benchmark::State& state);
@@ -75,27 +75,23 @@ class IndexBenchmark {
     std::uniform_int_distribution<> entity_id_distribution_;
 };
 
-template <dimension_t DIM>
-IndexBenchmark<DIM>::IndexBenchmark(
-    benchmark::State& state,
-    TestGenerator data_type,
-    int num_entities,
-    int updates_per_round,
-    double move_distance)
-: data_type_{data_type}
-, num_entities_(num_entities)
+template <dimension_t DIM, UpdateType UPDATE_TYPE>
+IndexBenchmark<DIM, UPDATE_TYPE>::IndexBenchmark(
+    benchmark::State& state, size_t updates_per_round, double move_distance)
+: data_type_{static_cast<TestGenerator>(state.range(1))}
+, num_entities_(state.range(0))
 , updates_per_round_(updates_per_round)
 , move_distance_(move_distance)
-, boxes_(num_entities)
+, boxes_(num_entities_)
 , updates_(updates_per_round)
 , random_engine_{0}
-, entity_id_distribution_{0, num_entities - 1} {
+, entity_id_distribution_{0, static_cast<int>(num_entities_ - 1)} {
     logging::SetupDefaultLogging();
     SetupWorld(state);
 }
 
-template <dimension_t DIM>
-void IndexBenchmark<DIM>::Benchmark(benchmark::State& state) {
+template <dimension_t DIM, UpdateType UPDATE_TYPE>
+void IndexBenchmark<DIM, UPDATE_TYPE>::Benchmark(benchmark::State& state) {
     for (auto _ : state) {
         state.PauseTiming();
         BuildUpdates();
@@ -105,8 +101,8 @@ void IndexBenchmark<DIM>::Benchmark(benchmark::State& state) {
     }
 }
 
-template <dimension_t DIM>
-void IndexBenchmark<DIM>::SetupWorld(benchmark::State& state) {
+template <dimension_t DIM, UpdateType UPDATE_TYPE>
+void IndexBenchmark<DIM, UPDATE_TYPE>::SetupWorld(benchmark::State& state) {
     logging::info("Setting up world with {} entities and {} dimensions.", num_entities_, DIM);
     CreateBoxData<DIM>(boxes_, data_type_, num_entities_, 0, GLOBAL_MAX, BOX_LEN);
     for (size_t i = 0; i < num_entities_; ++i) {
@@ -118,8 +114,8 @@ void IndexBenchmark<DIM>::SetupWorld(benchmark::State& state) {
     logging::info("World setup complete.");
 }
 
-template <dimension_t DIM>
-void IndexBenchmark<DIM>::BuildUpdates() {
+template <dimension_t DIM, UpdateType UPDATE_TYPE>
+void IndexBenchmark<DIM, UPDATE_TYPE>::BuildUpdates() {
     for (auto& update : updates_) {
         int box_id = entity_id_distribution_(random_engine_);
         update.id_ = box_id;
@@ -134,13 +130,36 @@ void IndexBenchmark<DIM>::BuildUpdates() {
 }
 
 template <dimension_t DIM>
-void IndexBenchmark<DIM>::UpdateWorld(benchmark::State& state) {
+size_t UpdateByRelocate(TreeType<DIM>& tree, std::vector<UpdateOp<DIM>>& updates) {
+    size_t n = 0;
+    for (auto& update : updates) {
+        n += tree.relocate(update.old_, update.new_);
+    }
+    return n;
+}
+
+template <dimension_t DIM>
+size_t UpdateByKey(TreeType<DIM>& tree, std::vector<UpdateOp<DIM>>& updates) {
+    size_t n = 0;
+    for (auto& update : updates) {
+        size_t result_erase = tree.erase(update.old_);
+        auto result_emplace = tree.emplace(update.new_, update.id_);
+        n += result_erase == 1 && result_emplace.second;
+    }
+    return n;
+}
+
+template <dimension_t DIM, UpdateType UPDATE_TYPE>
+void IndexBenchmark<DIM, UPDATE_TYPE>::UpdateWorld(benchmark::State& state) {
     size_t initial_tree_size = tree_.size();
     size_t n = 0;
-    for (auto& update : updates_) {
-        size_t result_erase = tree_.erase(update.old_);
-        auto result_emplace = tree_.emplace(update.new_, update.id_);
-        n += result_erase == 1 && result_emplace.second;
+    switch (UPDATE_TYPE) {
+    case UpdateType::ERASE_BY_KEY:
+        n = UpdateByKey(tree_, updates_);
+        break;
+    case UpdateType::RELOCATE:
+        n = UpdateByRelocate(tree_, updates_);
+        break;
     }
 
     if (n != updates_.size()) {
@@ -160,36 +179,28 @@ void IndexBenchmark<DIM>::UpdateWorld(benchmark::State& state) {
 }  // namespace
 
 template <typename... Arguments>
+void PhTreeRelocate3D(benchmark::State& state, Arguments&&... arguments) {
+    IndexBenchmark<3, UpdateType::RELOCATE> benchmark{state, arguments...};
+    benchmark.Benchmark(state);
+}
+
+template <typename... Arguments>
 void PhTree3D(benchmark::State& state, Arguments&&... arguments) {
-    IndexBenchmark<3> benchmark{state, arguments...};
+    IndexBenchmark<3, UpdateType::ERASE_BY_KEY> benchmark{state, arguments...};
     benchmark.Benchmark(state);
 }
 
 // index type, scenario name, data_type, num_entities, updates_per_round, move_distance
-// PhTree3D CUBE
-BENCHMARK_CAPTURE(PhTree3D, UPDATE_CU_100_of_1K, TestGenerator::CUBE, 1000)
+// PhTree with relocate()
+BENCHMARK_CAPTURE(PhTreeRelocate3D, UPDATE_1000, UPDATES_PER_ROUND)
+    ->RangeMultiplier(10)
+    ->Ranges({{1000, 1000 * 1000}, {TestGenerator::CUBE, TestGenerator::CLUSTER}})
     ->Unit(benchmark::kMillisecond);
 
-BENCHMARK_CAPTURE(PhTree3D, UPDATE_CU_100_of_10K, TestGenerator::CUBE, 10000)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, UPDATE_CU_100_of_100K, TestGenerator::CUBE, 100000)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, UPDATE_CU_100_of_1M, TestGenerator::CUBE, 1000000)
-    ->Unit(benchmark::kMillisecond);
-
-// PhTree3D CLUSTER
-BENCHMARK_CAPTURE(PhTree3D, UPDATE_CL_100_of_1K, TestGenerator::CLUSTER, 1000)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, UPDATE_CL_100_of_10K, TestGenerator::CLUSTER, 10000)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, UPDATE_CL_100_of_100K, TestGenerator::CLUSTER, 100000)
-    ->Unit(benchmark::kMillisecond);
-
-BENCHMARK_CAPTURE(PhTree3D, UPDATE_CL_100_of_1M, TestGenerator::CLUSTER, 1000000)
+// PhTree with relocate()
+BENCHMARK_CAPTURE(PhTree3D, UPDATE_1000, UPDATES_PER_ROUND)
+    ->RangeMultiplier(10)
+    ->Ranges({{1000, 1000 * 1000}, {TestGenerator::CUBE, TestGenerator::CLUSTER}})
     ->Unit(benchmark::kMillisecond);
 
 BENCHMARK_MAIN();