+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Added custom bucket implementation (similar to std::unordered_set). This improves update performance by 5%-20%.
   [#44](https://github.com/tzaeschke/phtree-cpp/issues/44)
+- Added `PhTree.relocate(old_key, new_key)` and `PhTree.relocate_if(old_key, new_key, predicate)`. 
+  This is **a lot faster** than using other methods. 
+  [#43](https://github.com/tzaeschke/phtree-cpp/issues/43)
 - Added try_emplace(key, value) and try_emplace(iter_hint, key, value)
   [#40](https://github.com/tzaeschke/phtree-cpp/issues/40)
 - Added FilterBoxAABB and FilterSphereAABB as examples for filtering a PH-Tree with box keys